requirements.txt
numpy>=1.20.0
scipy>=1.7.0
torch>=1.10.0
gymnasium>=0.26.0
plotly>=5.5.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0 